Completed Flex Exercise: Flex Layout 2
Added a container around the cards and another container around the sidebar and cards.
Removed text decoration and bulleting from the sidebar links.
Set the content class to wrap its content and to fill up availble space.
Side bar set to not shrink.
centered header text both vertically and horizontally.
footer pushed to the bottom of the page.